Add CLI with project creation and command generator
The idea behind this commit is to serve as a starting point for
a future where we can generate commands, queries, DEPs, and other
things via the commandline, similar to Rails' generators. This can
serve as a valuable starting point for people new to event sourcing,
as it helps get a basic directory structure up. Current usage is:

`eventsourcery new [PROJECT NAME]`
`eventsourcery generate:command [AGGREGATE] [COMMAND] [EVENT]`