Use configured database/tracker/event_(source|sink)
There are configuration settings for the projections database,
event tracker, event source, and event sink. This change makes it
so that if we wanted to, we don't have to pass in these dependencies
to ESPs (projectors and reactors)--they would use the configured
defaults instead. This should make setup of ESPs much less verbose
in the case that we only want the default of everything, which would
be most cases, while also not losing the flexibility that dependency
injection provides.