Add optional name parameter to table() method
Previously in projectors and reactors the `table()` method only
worked if there was exactly one table. This change maintains the
same functionality, where single tables can be accessed, but also
adds an _optional_ name parameter so that if there are multiple
tables, we can access them by specifying the name of the table.

In a projector/reactor, it will still be possible to do:

table.<operation> # works if only one table
table.<operation> # fails if (0 || >1) table

Additionally, it's now also possible to do:

table(:name).<operation> # works if :name is a table
table(:blah).<operation> # fails if :blah isn't a table