Use configured database by default in schema creation
This change makes it so that passing in a database becomes optional
when creating database tables. EventSourcery will now automatically
use EventSourcery.config.event_store if no database is passed in as
an argument. We have the database, so why not use it?

This also makes a _slight_ change to the schema creation API, by
making the database a kwarg.