add `table_prefix` attr_accessor to `TableOwner`
this allows projector (or other `TableOwner`) instances to specify a custom
naming prefix for all tables they own.

This is useful when multiple instances of a projector are running in parallel,
and each needs it's own set of working table.

An example of this is when using elasticsearch as a target for the projection
data as in elements-ingestion, and blue/green management of two indexes each
managed by it's own projector is used.