Resolve Sequel deprecation notice:
SEQUEL DEPRECATION WARNING: Calling a dataset filtering method with multiple
arguments or an array where the first argument/element is a string is deprecated
and will be removed in Sequel 5. Use Sequel.lit("id >= :from_id", {:from_id=>0})
to create an SQL fragment expression and pass that to the dataset filtering
method, or use the auto_literal_strings extension.

https://github.com/jeremyevans/sequel/commit/e795175e0d5a3063a815c771152e5f5cb05cf592